+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Allow to configure converse persistent store
 - Show a specific message to students when BBB meeting ended
 - Remove cancel button in student meeting join form
+- Replace automatic meeting joining by a button
 
 ## [4.0.0-beta.1] - 2022-02-15